remove boiler template, refactor file structure
