+++ src/components/components.jsx
@@ -1,0 +1,5 @@
+export { default as Footer } from './Footer/Footer';
+export { default as LoginForm } from './LoginForm/LoginForm';
+export { default as LogOutButton } from './LogOutButton/LogOutButton';
+export { default as Nav } from './Nav/Nav';
+export { default as RegisterForm } from './RegisterForm/RegisterForm';
