+++ src/index.jsx
@@ -4,8 +4,7 @@ import { Provider } from 'react-redux';
 
 import store from './redux/store';
 
-import App from './components/App/App';
-
+import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('react-root'));
 root.render(
