+++ src/pages/InfoPage/InfoPage.jsx
