+++ src/pages/pages.jsx
@@ -1,0 +1,6 @@
+export { default as AboutPage } from './AboutPage/AboutPage';
+export { default as InfoPage } from './InfoPage/InfoPage';
+export { default as LandingPage } from './LandingPage/LandingPage';
+export { default as LoginPage } from './LoginPage/LoginPage';
+export { default as RegisterPage } from './RegisterPage/RegisterPage';
+export { default as UserPage } from './UserPage/UserPage';
