+++ src/route/ProtectedRoute/ProtectedRoute.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Redirect } from 'react-router-dom';
-import LoginPage from '../LoginPage/LoginPage';
-import {useSelector} from 'react-redux';
+import { LoginPage } from '../../pages/pages';
+import { useSelector } from 'react-redux';
 
 // A Custom Wrapper Component -- This will keep our code DRY.
 // Responsible for watching redux state, and returning an appropriate component
@@ -27,15 +27,14 @@ function ProtectedRoute({ component, children, ...props }) {
       // are now passed along to the 'Route' Component
       {...props}
     >
-      {user.id ?
+      {user.id ? (
         // If the user is logged in, show the protected component
         <ProtectedComponent />
-        :
+      ) : (
         // Otherwise, redirect to the Loginpage
         <LoginPage />
-      }
+      )}
     </Route>
-
   );
 }
